clinica/siteclinica/requirements.txt
asgiref==3.10.0
cx-Oracle==8.3.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
mssql-django==1.6
mysqlclient==2.2.7
psycopg2-binary==2.9.11
pyodbc==5.3.0
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.15.0
